Copying Original Repository
This is a copy of the orginal repository used for this project as the original is owned by the university. In order to run and or port this project you will need unreal engine 5, the android SDK and ARCore installed. Howver all the code and assets created for this project should be accessible without those programs.

Cameron